technical task/backend/requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
huggingface_hub